requirements.txt
feedparser==6.0.11
Jinja2==3.1.4
python-dateutil==2.9.0.post0

# Upwork scraper dependencies
playwright>=1.44.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
